automation/requirements.txt
requests==2.26.0
ruamel.yaml==0.17.16
python-dateutil==2.8.2
loguru==0.5.3
readme-renderer==28.0
python-dotenv==0.19.0
click==8.0.1
colorteller